requirements.txt
langchain>=0.2.0
langgraph>=0.0.60
langchain_community>=0.2.0
langchain_core>=0.2.0
ollama>=0.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pypdf>=4.0.0
python-docx>=1.1.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiofiles>=23.0.0
